+++ zipline/pipeline/data/dividends.py
@@ -7,15 +7,15 @@ from .dataset import Column, DataSet
 
 
 class DividendsByExDate(DataSet):
-    next_ex_date = Column(datetime64ns_dtype)
-    previous_ex_date = Column(datetime64ns_dtype)
+    next_date = Column(datetime64ns_dtype)
+    previous_date = Column(datetime64ns_dtype)
     next_amount = Column(float64_dtype)
     previous_amount = Column(float64_dtype)
 
 
 class DividendsByPayDate(DataSet):
-    next_pay_date = Column(datetime64ns_dtype)
-    previous_pay_date = Column(datetime64ns_dtype)
+    next_date = Column(datetime64ns_dtype)
+    previous_date = Column(datetime64ns_dtype)
     next_amount = Column(float64_dtype)
     previous_amount = Column(float64_dtype)
