+++ zipline/pipeline/loaders/blaze/buyback_auth.py
@@ -68,24 +68,7 @@ class BlazeCashBuybackAuthorizationsLoader(BlazeEventsLoader):
     })
 
     concrete_loader = CashBuybackAuthorizationsLoader
-
-    def __init__(self,
-                 expr,
-                 resources=None,
-                 odo_kwargs=None,
-                 data_query_time=None,
-                 data_query_tz=None,
-                 dataset=CashBuybackAuthorizations,
-                 **kwargs):
-        super(
-            BlazeCashBuybackAuthorizationsLoader, self
-        ).__init__(expr,
-                   resources=resources,
-                   odo_kwargs=odo_kwargs,
-                   data_query_time=data_query_time,
-                   data_query_tz=data_query_tz,
-                   dataset=dataset,
-                   **kwargs)
+    default_dataset = CashBuybackAuthorizations
 
 
 class BlazeShareBuybackAuthorizationsLoader(BlazeEventsLoader):
@@ -140,21 +123,4 @@ class BlazeShareBuybackAuthorizationsLoader(BlazeEventsLoader):
     })
 
     concrete_loader = ShareBuybackAuthorizationsLoader
-
-    def __init__(self,
-                 expr,
-                 resources=None,
-                 odo_kwargs=None,
-                 data_query_time=None,
-                 data_query_tz=None,
-                 dataset=ShareBuybackAuthorizations,
-                 **kwargs):
-        super(
-            BlazeShareBuybackAuthorizationsLoader, self
-        ).__init__(expr,
-                   resources=resources,
-                   odo_kwargs=odo_kwargs,
-                   data_query_time=data_query_time,
-                   data_query_tz=data_query_tz,
-                   dataset=dataset,
-                   **kwargs)
+    default_dataset = ShareBuybackAuthorizations
