+++ zipline/pipeline/loaders/dividends.py
@@ -52,14 +52,14 @@ class DividendsByPayDateLoader(EventsLoader):
         )
 
     @lazyval
-    def next_pay_date_loader(self):
-        return self._next_event_date_loader(self.dataset.next_pay_date,
+    def next_date_loader(self):
+        return self._next_event_date_loader(self.dataset.next_date,
                                             PAY_DATE_FIELD_NAME)
 
     @lazyval
-    def previous_pay_date_loader(self):
+    def previous_date_loader(self):
         return self._previous_event_date_loader(
-            self.dataset.previous_pay_date,
+            self.dataset.previous_date,
             PAY_DATE_FIELD_NAME
         )
 
@@ -90,14 +90,14 @@ class DividendsByExDateLoader(EventsLoader):
         )
 
     @lazyval
-    def next_ex_date_loader(self):
-        return self._next_event_date_loader(self.dataset.next_ex_date,
+    def next_date_loader(self):
+        return self._next_event_date_loader(self.dataset.next_date,
                                             EX_DATE_FIELD_NAME)
 
     @lazyval
-    def previous_ex_date_loader(self):
+    def previous_date_loader(self):
         return self._previous_event_date_loader(
-            self.dataset.previous_ex_date,
+            self.dataset.previous_date,
             EX_DATE_FIELD_NAME
         )
